airflow/requirements.txt
pandas==2.2.2
numpy==1.26.4
SQLAlchemy==1.4.52
# SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
scikit-learn==1.4.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
seaborn==0.13.2
notebook==7.1.3
statsmodels==0.14.2
dbt-postgres==1.8.0
apache-airflow-providers-celery>=3.3.0
python-dotenv==1.0.1